基础功能优化 & 新颜色

+++ D4/WriteView+Rects.swift
@@ -23,23 +23,23 @@ extension WriteView {
 		return CGRectMake(0, ScreenHeight - 20, ScreenWidth, 20)
 	}
 
-	func blockFrame(index: CGFloat) -> CGRect {
+	func blockFrame(index: Int) -> CGRect {
 		let addend = index == 0 ? 0 : smallBlockHeight
-		let factor = index == 0 ? 0 : index - 1
+		let factor = index == 0 ? 0 : CGFloat(index - 1)
 		let y = addend + bigBlockHeight * factor
 		let height = (index == 0 || index == 4) ? smallBlockHeight : bigBlockHeight
 		return CGRectMake(0, y, ScreenWidth, height)
 	}
 
-	func activateBlockFrame(index: CGFloat) -> CGRect {
+	func activateBlockFrame(index: Int) -> CGRect {
 		var rect = blockFrame(index)
 		rect.size.width -= 100
 		return rect
 	}
 
-	func dotFrame(index: CGFloat) -> CGRect {
+	func dotFrame(index: Int) -> CGRect {
 		let x = ScreenWidth - 50
-		let y = bigBlockHeight * index
+		let y = bigBlockHeight * CGFloat(index)
 		return CGRectMake(x, y, dotLength, dotLength)
 	}
 }
